Move functions to better match reponsibility
